factored out custom fetch response function into session server
